+++ cli/tsc/diagnostics.rs
@@ -29,15 +29,6 @@ const UNSTABLE_DENO_PROPS: &[&str] = &[
   "removeSignalListener",
   "shutdown",
   "umask",
-  "Child",
-  "ChildProcess",
-  "ChildStatus",
-  "SpawnOutput",
-  "command",
-  "Command",
-  "CommandOptions",
-  "CommandStatus",
-  "CommandOutput",
   "serve",
   "ServeInit",
   "ServeTlsInit",
