+++ runtime/js/99_main.js
@@ -62,7 +62,6 @@ import { errors } from "internal:runtime/js/01_errors.js";
 import * as webidl from "internal:deno_webidl/00_webidl.js";
 import DOMException from "internal:deno_web/01_dom_exception.js";
 import * as flash from "internal:deno_flash/01_http.js";
-import * as spawn from "internal:runtime/js/40_spawn.js";
 import {
   mainRuntimeGlobalProperties,
   setLanguage,
@@ -466,15 +465,6 @@ function bootstrapMainRuntime(runtimeOptions) {
   // a snapshot
   ObjectAssign(internals, {
     nodeUnstable: {
-      Command: spawn.createCommand(
-        spawn.createSpawn(ops.op_node_unstable_spawn_child),
-        spawn.createSpawnSync(
-          ops.op_node_unstable_spawn_sync,
-        ),
-        spawn.createSpawnChild(
-          ops.op_node_unstable_spawn_child,
-        ),
-      ),
       serve: flash.createServe(ops.op_node_unstable_flash_serve),
       upgradeHttpRaw: flash.upgradeHttpRaw,
       listenDatagram: net.createListenDatagram(
@@ -513,11 +503,6 @@ function bootstrapMainRuntime(runtimeOptions) {
     // the op function that needs to be passed will be invalidated by creating
     // a snapshot
     ObjectAssign(finalDenoNs, {
-      Command: spawn.createCommand(
-        spawn.createSpawn(ops.op_spawn_child),
-        spawn.createSpawnSync(ops.op_spawn_sync),
-        spawn.createSpawnChild(ops.op_spawn_child),
-      ),
       serve: flash.createServe(ops.op_flash_serve),
       listenDatagram: net.createListenDatagram(
         ops.op_net_listen_udp,
@@ -611,15 +596,6 @@ function bootstrapWorkerRuntime(
   // a snapshot
   ObjectAssign(internals, {
     nodeUnstable: {
-      Command: spawn.createCommand(
-        spawn.createSpawn(ops.op_node_unstable_spawn_child),
-        spawn.createSpawnSync(
-          ops.op_node_unstable_spawn_sync,
-        ),
-        spawn.createSpawnChild(
-          ops.op_node_unstable_spawn_child,
-        ),
-      ),
       serve: flash.createServe(ops.op_node_unstable_flash_serve),
       upgradeHttpRaw: flash.upgradeHttpRaw,
       listenDatagram: net.createListenDatagram(
@@ -650,11 +626,6 @@ function bootstrapWorkerRuntime(
     // the op function that needs to be passed will be invalidated by creating
     // a snapshot
     ObjectAssign(finalDenoNs, {
-      Command: spawn.createCommand(
-        spawn.createSpawn(ops.op_spawn_child),
-        spawn.createSpawnSync(ops.op_spawn_sync),
-        spawn.createSpawnChild(ops.op_spawn_child),
-      ),
       serve: flash.createServe(ops.op_flash_serve),
       listenDatagram: net.createListenDatagram(
         ops.op_net_listen_udp,
